Red team code, + a few ideas
The biggest issue I've found after a cursory glance is that you can vote
for multiple candidates IF you don't vote for the first candidate. Maybe
they assumed there could only be 2 candidates.
The other major issue is that they don't generate random primes for
their crypto, so we could leak the ctxt and bruteforce it to get the
original values if we wanted to.
Another thing that is interesting but not a major bug is their parsing
of voters.txt: empty strings are valid. So if your input file ends with
a newline (like it probably does), there is secretly a voter named empty
string. The first person to realize this gets to vote as them.